+++ extension-files/bringweb3-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bringweb3/chrome-extension-kit
 
+## 1.6.2
+
+### Patch Changes
+
+- 24ef774: Removed the automatic change from activated to quiet.
+
+## 1.6.1
+
+### Patch Changes
+
+- 692ce76: quietDomains types handling, iframe styles src
+
 ## 1.6.0
 
 ### Minor Changes